+++ src/main/java/net/silentchaos512/mechanisms/item/CraftingItems.java
@@ -11,17 +11,20 @@ import net.silentchaos512.utils.TextUtils;
 import javax.annotation.Nullable;
 
 public enum CraftingItems implements IItemProvider {
+    // Crafting items
+    CIRCUIT_BOARD,
+    // Mod metals
     REDSTONE_ALLOY_INGOT,
     REDSTONE_ALLOY_NUGGET,
     REDSTONE_ALLOY_DUST,
     REFINED_IRON_INGOT,
     COMPRESSED_IRON_INGOT,
-    CIRCUIT_BOARD,
+    // Base metals
+    COPPER_CHUNKS,
     IRON_CHUNKS,
     IRON_DUST,
     GOLD_CHUNKS,
     GOLD_DUST,
-    COPPER_CHUNKS,
     COPPER_DUST,
     COPPER_INGOT,
     TIN_CHUNKS,
@@ -36,10 +39,22 @@ public enum CraftingItems implements IItemProvider {
     NICKEL_CHUNKS,
     NICKEL_DUST,
     NICKEL_INGOT,
-    BRONZE_INGOT,
+    ZINC_CHUNKS,
+    ZINC_DUST,
+    ZINC_INGOT,
+    BISMUTH_CHUNKS,
+    BISMUTH_DUST,
+    BISMUTH_INGOT,
+    // Alloys
     BRONZE_DUST,
-    INVAR_INGOT,
+    BRONZE_INGOT,
+    BRASS_DUST,
+    BRASS_INGOT,
+    BISMUTH_BRASS_DUST,
+    BISMUTH_BRASS_INGOT,
     INVAR_DUST,
+    INVAR_INGOT,
+    // Others
     ZOMBIE_LEATHER,
     BEEF_JERKY(Foods.COOKED_BEEF),
     PORK_JERKY(Foods.COOKED_PORKCHOP),