+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Zinc, bismuth, brass, and bismuth brass
+
 ## [0.3.5] - 2019-08-19
 ### Added
 - Missing crushing recipes for silver and lead